Reduce code duplication through restructuring.
