+++ tests/quasi-newton/helpers.cpp
