+++ tests/quasi-newton/helpers.hpp
