+++ tests/quasi-newton/serial/TestQN2.cpp
@@ -3,20 +3,20 @@
 #include "testing/Testing.hpp"
 
 #include <precice/SolverInterface.hpp>
-#include "helpers.hpp"
+#include "../helpers.hpp"
 
 BOOST_AUTO_TEST_SUITE(Integration)
-BOOST_AUTO_TEST_SUITE(Serial)
 BOOST_AUTO_TEST_SUITE(QuasiNewton)
-BOOST_AUTO_TEST_CASE(TestQN1)
+BOOST_AUTO_TEST_SUITE(Serial)
+BOOST_AUTO_TEST_CASE(TestQN2)
 {
   PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-  // serial coupling, IQN-ILS, strict QR2 filter
-  tests::serial::runTestQN(context.config(), context);
+  // parallel coupling, IQN-ILS, strict QR2 filter
+  runTestQN(context.config(), context);
 }
 
 BOOST_AUTO_TEST_SUITE_END() // Integration
-BOOST_AUTO_TEST_SUITE_END() // Serial
 BOOST_AUTO_TEST_SUITE_END() // QuasiNewton
+BOOST_AUTO_TEST_SUITE_END() // Serial
 
 #endif // PRECICE_NO_MPI
