+++ tests/quasi-newton/serial/TestQN3.cpp
@@ -3,20 +3,20 @@
 #include "testing/Testing.hpp"
 
 #include <precice/SolverInterface.hpp>
-#include "helpers.hpp"
+#include "../helpers.hpp"
 
 BOOST_AUTO_TEST_SUITE(Integration)
-BOOST_AUTO_TEST_SUITE(Serial)
 BOOST_AUTO_TEST_SUITE(QuasiNewton)
-BOOST_AUTO_TEST_CASE(TestQN2)
+BOOST_AUTO_TEST_SUITE(Serial)
+BOOST_AUTO_TEST_CASE(TestQN3)
 {
   PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-  // parallel coupling, IQN-ILS, strict QR2 filter
-  tests::serial::runTestQN(context.config(), context);
+  // serial coupling, IQN-IMVJ (which is identical to IQN-ILS as only first timestep is considered), relaxed QR2 filter
+  runTestQN(context.config(), context);
 }
 
 BOOST_AUTO_TEST_SUITE_END() // Integration
-BOOST_AUTO_TEST_SUITE_END() // Serial
 BOOST_AUTO_TEST_SUITE_END() // QuasiNewton
+BOOST_AUTO_TEST_SUITE_END() // Serial
 
 #endif // PRECICE_NO_MPI
